\name{packageSkeletonForR-package}
\alias{packageSkeletonForR-package}
\docType{package}
\title{A GitHub based package skeleton for R together with a static website
}
\description{
A GitHub based package skeleton for R. The repo includes the package skeleton, yml and markdown to produce a static website using additional R packages (packagePages and centerPackageTemplate)
}

\details{
\tabular{ll}{
Package: \tab packageSkeletonForR\cr
Type: \tab Package\cr
Version: \tab 0.0-0.1\cr
Date: \tab 2021-06-21\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}

\author{Damian W. Betebenner \email{DBetebenner@nciea.org}}

\keyword{package}
